TASK 1/requirements.txt
anyio  ==  4.1.0
argon2-cffi ==   23.1.0
argon2-cffi-bindings==   21.2.0
arrow==  1.3.0
asttokens==  2.4.1
async-lru ==    2.0.4
attrs ==   23.1.0
Babel==   2.14.0
beautifulsoup4==  4.12.2
bleach ==   6.1.0
certifi==   2022.12.7
cffi==      1.16.0
charset-normalizer        2.1.1
colorama==  0.4.6
comm==      0.1.4
contourpy== 1.2.0
cycler==    0.12.1
debugpy==   1.6.7
decorator== 5.1.1
defusedxml==0.7.1
exceptiongroup== 1.2.0
executing== 2.0.1
fastjsonschema== 2.19.0
filelock==  3.9.0
fonttools== 4.46.0
fqdn== 1.5.1
fsspec== 2023.4.0
idna== 3.4
importlib-metadata==7.0.0
ipykernel== 6.26.0
ipython== 8.18.1
isoduration == 20.11.0
jedi== 0.19.1
Jinja2== 3.1.2
json5==  0.9.14
jsonpointer==  2.4
jsonschema==4.20.0
jsonschema-specifications 2023.11.2
jupyter_client==  8.6.0
jupyter_core== 5.5.0
jupyter-events ==0.9.0
jupyter-lsp==    2.2.1
jupyter_server  == 2.12.1
jupyter_server_terminals == 0.5.0
jupyterlab==4.0.9
jupyterlab_pygments == 0.3.0
jupyterlab_server == 2.25.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==   3.0.2
mpmath==    1.3.0
nbclient==  0.9.0
nbconvert== 7.12.0
nbformat==  5.9.2
nest-asyncio==1.5.8
networkx==  3.0
notebook==  7.0.6
notebook_shim== 0.2.3
numpy== 1.24.1
opencv-python==4.8.1.78
overrides== 7.4.0
packaging== 23.2
pandas== 2.1.4
pandocfilters==1.5.0
parso==     0.8.3
pickleshare == 0.7.5
Pillow== 9.3.0
pip==  23.3.1
platformdirs==4.1.0
prometheus-client== 0.19.0
prompt-toolkit == 3.0.41
psutil==    5.9.0
pure-eval== 0.2.2
py-cpuinfo==9.0.0
pycparser== 2.21
Pygments==  2.17.2
pyparsing== 3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==      2023.3.post1
pywin32==   305.1
pywinpty==  2.0.12
PyYAML==    6.0.1
pyzmq==     25.1.2
referencing==  0.32.0
requests==  2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==   0.13.2
scipy==     1.11.4
seaborn==   0.13.0
Send2Trash==1.8.2
setuptools==68.0.0
six==       1.16.0
sniffio==   1.3.0
soupsieve== 2.5
stack-data==0.6.2
sympy==     1.12
terminado== 0.18.0
thop==      0.1.1.post2209072238
tinycss2==  1.2.1
tomli==     2.0.1
torch==     2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision== 0.16.0+cu121
tornado==   6.2
tqdm==  4.66.1
traitlets== 5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.4.0
tzdata== 2023.3
ultralytics==8.0.227
uri-template==1.3.0
urllib3==   1.26.13
wcwidth==   0.2.12
webcolors== 1.13
webencodings==0.5.1
websocket-client== 1.7.0
wheel== 0.41.2
zipp==3.17.0